Fix CDN and download tests (#68)
* (chore) Initial fixes to make CDN tests pass eventually

By default they will run on Prod and fail until
www.firefox.com is up, so I've disabled the scheduled test
for now.

But here we can try them on Dev or Stage manually in the
meantime.

* (chore) Disable scheduled Download tests for now

They're intended to run on www.firefox.com (Prod) but that
is not yet up, so disabling for now. Can run them manually
on Dev or Stage to check they work