Fix link to appropriate version of ESR releasenotes (#280)
This is because we now have more than one ESR available at any one time
and the current logic doesn't expect that so just finds the youngest ESR record.

We use the product_details lib as the source of truth for which version
we want for the "esr" channel, because we can't depend on the ESR
versions being saved/created in a reliable order.

Manually testable by clicking the 'Release notes' link on http://localhost:8000/en-US/download/all/desktop-esr/

Port of https://github.com/mozilla/bedrock/pull/16328/