Include Firefox fonts and theming on WNP
This is restricted to WNP to keep review scope small. We currently
have custom property overrides for the new Mozilla theme. These are
further overridden for Firefox theme class link values.

Given that we didn't include the Firefox fonts and theme overrides
on global protocol stylesheet, we're keeping them scoped to WNP for
this changeset.