Add evergreen WNPs (#441)
* Add basic WNP logic and Nightly WNP

* Add developer WNP

* Add default evergreen WNP

* Remove unnecessary email block

* Remove unused base ftl from Nightly WNP

* Simplify base evergreen FTL

* Remove unused variant QR codes (experiment is finished)

Note: the image should be replaced with qr_code helper in future

* Add tests, fix ruff lint error

* Remove legacy logic around pre-75 dev WNP template

* Include Firefox fonts and theming on WNP

This is restricted to WNP to keep review scope small. We currently
have custom property overrides for the new Mozilla theme. These are
further overridden for Firefox theme class link values.

Given that we didn't include the Firefox fonts and theme overrides
on global protocol stylesheet, we're keeping them scoped to WNP for
this changeset.

* Remove use of third party brand strings

* Remove unused release WNP JS

* Add WNP Fx theme title sizes

* Add whatsnew FTL files to pontoon config