Pattern Library Updates (#695)
* Update wnp page for pattern library

* Add blog template WIP

* Add flare components docs stub

* Update flare component docs

* Fix toc

* Improve flare components documentation

* Rename pattern library pages directories

* Fix component inconsistencies

This commit makes all flare component files and CSS consistent with the documentation, but it has a broad scope and might cause merge conflicts, so we might want to drop this commit and come back to this later.

* Add markdown component

* Add docs and blog markdown

* Rename markdown to rich_text

This is actually a rich_text component, not markdown. Markdown is only really used in the pattern_library. Wagtail doesn't support markdown input.

* Update blog page layout

* Enable markdown toc links

* Address copilot feedback

---------

Co-authored-by: Mariana Bedran Lesche <maribedran@gmail.com>