(fix) Run the static-assets check from within the relevant container
This changeset amends the workflow so that we don't need to install Django
in the runner to trigger the static-assets check. It now exists in the image,
so we can use it there, passing it the path to the staticfiles.json that
should also exist in the prebuilt container. (If it does not, for some reason
the check will rightly fail)